we can save a bit of time by pushing api key updates to after the http response is sent
using Datadog traces I was able to determine we spend the last 20-30ms of every API request (using API keys) simply updating the last time the API key was accessed (db write).

I confirmed with our devops team that we can support a terminable middleware and this moves the database write to after the HTTP response is sent to the client. This should provide a modest speed improvement.